Make parse_hdr a little more robust
Turns out, according to the SBE manuals I can find, these headers can be anything after a * at the top of the file.
This change removes the assumption that only ** items are comments and instead picks out some information:
* data file
* software version

Then assumes that everything without an equal sign (=) is a comment.
Everything with an equal sign is split into a key:value pair