handle mixed line ending header section of hex files
R2R in their routine processing, received a set of files that has "mixed" line
endings in the comments section of a hex file. This resulted in behavior where
blank lines would appear in the comments section. As written, these would be
interpreted as zero length data lines and cause some pretty invalid scan
numbers to be reported.

This fix switches the line splitting to be on the sequence \r\n and leaves any
duplicate \r chars intact. It also switches the to_hex accessor method to use
split("\n") to account for the same situation. Hopefully this does not break
compatibility with existing netCDF files that have been converted from hex.